All int types changed - B10MoveStrategy

+++ blackJack_Simulation/blackJack_Simulation/B10BetStrategy.cs
@@ -2,9 +2,9 @@
 {
     class B10BetStrategy : IBetStrategy
     {
-        private double[] cardAmounts;
+        private int[] cardAmounts;
 
-        public B10BetStrategy(double[] cardAmounts)
+        public B10BetStrategy(int[] cardAmounts)
         {
             this.cardAmounts = cardAmounts;
         }
